[#21,view/map][m]: add hover-over labels to items on the map using Leaflet.Label plugin [1].
[1]: https://github.com/Leaflet/Leaflet.label

NB: I spent about 1h trying to work out how to have labels show always but it
seems this is really hard given our setup (where we don't know when markers are
added to the map ...). What we need to do is iterate through markers *after*
they are added to map and called showLabel but I could not work out how to do
this.